+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/transaction/impl/SnapshotTransaction.java
@@ -38,6 +38,9 @@ import org.apache.commons.lang3.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.codahale.metrics.Histogram;
+import com.codahale.metrics.MetricRegistry;
+import com.codahale.metrics.Timer;
 import com.google.common.base.Function;
 import com.google.common.base.Functions;
 import com.google.common.base.MoreObjects;
@@ -101,6 +104,7 @@ import com.palantir.atlasdb.transaction.api.TransactionFailedRetriableException;
 import com.palantir.atlasdb.transaction.api.TransactionLockTimeoutException;
 import com.palantir.atlasdb.transaction.api.TransactionReadSentinelBehavior;
 import com.palantir.atlasdb.transaction.service.TransactionService;
+import com.palantir.atlasdb.util.AtlasDbMetrics;
 import com.palantir.common.annotation.Output;
 import com.palantir.common.base.AbortingVisitor;
 import com.palantir.common.base.AbstractBatchingVisitable;
@@ -189,6 +193,9 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
     protected final boolean allowHiddenTableAccess;
     protected final Stopwatch transactionTimer = Stopwatch.createStarted();
     protected final TimestampCache timestampValidationReadCache;
+
+    private final MetricRegistry metricRegistry = AtlasDbMetrics.getMetricRegistry();
+    private final Timer.Context transactionTimerContext = getTimer("transactionMillis").time();
 
     /**
      * @param immutableTimestamp If we find a row written before the immutableTimestamp we don't need to
@@ -327,7 +334,7 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
     @Override
     public SortedMap<byte[], RowResult<byte[]>> getRows(TableReference tableRef, Iterable<byte[]> rows,
                                                         ColumnSelection columnSelection) {
-        Stopwatch watch = Stopwatch.createStarted();
+        Timer.Context timer = getTimer("getRows").time();
         checkGetPreconditions(tableRef);
         if (Iterables.isEmpty(rows)) {
             return AbstractTransaction.EMPTY_SORTED_ROWS;
@@ -346,9 +353,10 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
         rawResults.keySet().removeAll(result.keySet());
 
         SortedMap<byte[], RowResult<byte[]>> results = filterRowResults(tableRef, rawResults, result);
+        long getRowsMillis = TimeUnit.NANOSECONDS.toMillis(timer.stop());
         if (perfLogger.isDebugEnabled()) {
             perfLogger.debug("getRows({}, {} rows) found {} rows, took {} ms",
-                    tableRef, Iterables.size(rows), results.size(), watch.elapsed(TimeUnit.MILLISECONDS));
+                    tableRef, Iterables.size(rows), results.size(), getRowsMillis);
         }
         validateExternalAndCommitLocksIfNecessary(tableRef);
         return results;
@@ -545,7 +553,7 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
 
     @Override
     public Map<Cell, byte[]> get(TableReference tableRef, Set<Cell> cells) {
-        Stopwatch watch = Stopwatch.createStarted();
+        Timer.Context timer = getTimer("get").time();
         checkGetPreconditions(tableRef);
         if (Iterables.isEmpty(cells)) {
             return ImmutableMap.of();
@@ -564,9 +572,10 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
         // We don't need to read any cells that were written locally.
         result.putAll(getFromKeyValueService(tableRef, Sets.difference(cells, result.keySet())));
 
+        long getMillis = TimeUnit.NANOSECONDS.toMillis(timer.stop());
         if (perfLogger.isDebugEnabled()) {
             perfLogger.debug("get({}, {} cells) found {} cells (some possibly deleted), took {} ms",
-                    tableRef, cells.size(), result.size(), watch.elapsed(TimeUnit.MILLISECONDS));
+                    tableRef, cells.size(), result.size(), getMillis);
         }
         validateExternalAndCommitLocksIfNecessary(tableRef);
         return Maps.filterValues(result, Predicates.not(Value.IS_EMPTY));
@@ -621,7 +630,7 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
                 .transformAndConcat(new Function<List<RangeRequest>, List<BatchingVisitable<RowResult<byte[]>>>>() {
                     @Override
                     public List<BatchingVisitable<RowResult<byte[]>>> apply(List<RangeRequest> input) {
-                        Stopwatch timer = Stopwatch.createStarted();
+                        Timer.Context timer = getTimer("processedRangeMillis").time();
                         Map<RangeRequest, TokenBackedBasicResultsPage<RowResult<Value>, byte[]>> firstPages =
                                 keyValueService.getFirstBatchForRanges(tableRef, input, getStartTimestamp());
                         validateExternalAndCommitLocksIfNecessary(tableRef);
@@ -667,8 +676,9 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
                                 }
                             });
                         }
+                        long processedRangeMillis = TimeUnit.NANOSECONDS.toMillis(timer.stop());
                         log.trace("Processed {} range requests for {} in {}ms",
-                                input.size(), tableRef, timer.elapsed(TimeUnit.MILLISECONDS));
+                                input.size(), tableRef, processedRangeMillis);
                         return ret;
                     }
 
@@ -1186,9 +1196,10 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
 
             checkConstraints();
             commitWrites(transactionService);
+            long transactionMillis = TimeUnit.NANOSECONDS.toMillis(transactionTimerContext.stop());
             perfLogger.debug("Committed transaction {} in {}ms",
                     getStartTimestamp(),
-                    getTrasactionTimer().elapsed(TimeUnit.MILLISECONDS));
+                    transactionMillis);
             success = true;
         } finally {
             // Once we are in state committing, we need to try/finally to set the state to a terminal state.
@@ -1218,17 +1229,17 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
         if (!hasWrites()) {
             return;
         }
-        Stopwatch watch = Stopwatch.createStarted();
-        LockRefreshToken commitLocksToken = acquireLocksForCommit();
-        long millisForLocks = watch.elapsed(TimeUnit.MILLISECONDS);
-        try {
-            watch.reset().start();
-            throwIfConflictOnCommit(commitLocksToken, transactionService);
-            long millisCheckingForConflicts = watch.elapsed(TimeUnit.MILLISECONDS);
 
-            watch.reset().start();
+        Timer.Context acquireLocksTimer = getTimer("commitAcquireLocks").time();
+        LockRefreshToken commitLocksToken = acquireLocksForCommit();
+        long millisForLocks = TimeUnit.NANOSECONDS.toMillis(acquireLocksTimer.stop());
+        try {
+            Timer.Context conflictsTimer = getTimer("commitCheckingForConflicts").time();
+            throwIfConflictOnCommit(commitLocksToken, transactionService);
+            long millisCheckingForConflicts = TimeUnit.NANOSECONDS.toMillis(conflictsTimer.stop());
+            Timer.Context writesTimer = getTimer("commitWrite").time();
             keyValueService.multiPut(writesByTable, getStartTimestamp());
-            long millisForWrites = watch.elapsed(TimeUnit.MILLISECONDS);
+            long millisForWrites = TimeUnit.NANOSECONDS.toMillis(writesTimer.stop());
 
             // Now that all writes are done, get the commit timestamp
             // We must do this before we check that our locks are still valid to ensure that
@@ -1240,9 +1251,9 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
             // punch on commit so that if hard delete is the only thing happening on a system,
             // we won't block forever waiting for the unreadable timestamp to advance past the
             // scrub timestamp (same as the hard delete transaction's start timestamp)
-            watch.reset().start();
+            Timer.Context punchTimer = getTimer("millisForPunch").time();
             cleaner.punch(commitTimestamp);
-            long millisForPunch = watch.elapsed(TimeUnit.MILLISECONDS);
+            long millisForPunch = TimeUnit.NANOSECONDS.toMillis(punchTimer.stop());
 
             throwIfReadWriteConflictForSerializable(commitTimestamp);
 
@@ -1250,9 +1261,9 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
             // this check is required by the transaction protocol for correctness
             throwIfExternalAndCommitLocksNotValid(commitLocksToken);
 
-            watch.reset().start();
+            Timer.Context commitTsTimer = getTimer("commitPutCommitTs").time();
             putCommitTimestamp(commitTimestamp, commitLocksToken, transactionService);
-            long millisForCommitTs = watch.elapsed(TimeUnit.MILLISECONDS);
+            long millisForCommitTs = TimeUnit.NANOSECONDS.toMillis(commitTsTimer.stop());
 
             Set<LockRefreshToken> expiredLocks = refreshExternalAndCommitLocks(commitLocksToken);
             if (!expiredLocks.isEmpty()) {
@@ -1263,6 +1274,9 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
                 log.error(errorMessage, new TransactionFailedRetriableException(errorMessage));
             }
             long millisSinceCreation = System.currentTimeMillis() - timeCreated;
+            getTimer("commitTotalTimeSinceTxCreation").update(millisSinceCreation, TimeUnit.MILLISECONDS);
+            Histogram byteSizeTx = getHistogram("byteSizeTx");
+            byteSizeTx.update(byteCount.get());
             if (perfLogger.isDebugEnabled()) {
                 perfLogger.debug("Committed {} bytes with locks, start ts {}, commit ts {}, "
                         + "acquiring locks took {} ms, checking for conflicts took {} ms, "
@@ -1700,10 +1714,11 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
 
         // Before we do the reads, we need to make sure the committer is done writing.
         if (waitForCommitterToComplete) {
-            Stopwatch watch = Stopwatch.createStarted();
+            Timer.Context timer = getTimer("waitForCommitTsMillis").time();
             waitForCommitToComplete(startTimestamps);
+            long waitForCommitTsMillis = TimeUnit.NANOSECONDS.toMillis(timer.stop());
             perfLogger.debug("Waited {} ms to get commit timestamps for table {}.",
-                    watch.elapsed(TimeUnit.MILLISECONDS), tableRef);
+                    waitForCommitTsMillis, tableRef);
         }
 
         log.trace("Getting commit timestamps for {} start timestamps in response to read from table {}",
@@ -1841,6 +1856,15 @@ public class SnapshotTransaction extends AbstractTransaction implements Constrai
         return tableRefToCells;
     }
 
+    private Timer getTimer(String name) {
+        return metricRegistry.getTimers()
+                .getOrDefault(name, metricRegistry.timer(MetricRegistry.name(SnapshotTransaction.class, name)));
+    }
+
+    private Histogram getHistogram(String name) {
+        return metricRegistry.getHistograms()
+                .getOrDefault(name, metricRegistry.histogram(MetricRegistry.name(SnapshotTransaction.class, name)));
+    }
 }